# docs/swagger.yaml
openapi: 3.0.0
info:
  title: CarLedgr API
  description: |
    Comprehensive dealership management system API for vehicle repair and resale businesses.
    
    **Authentication:** Most endpoints require JWT token in Authorization header: `Bearer <token>`
    
    **Organization Isolation:** All data is automatically scoped to the authenticated user's organization
    
    **License Enforcement:** Some endpoints check subscription status and car limits
  version: 2.0.0
  contact:
    name: CarLedgr API Support
    url: https://carledgr.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3030/api
    description: Development server
  - url: https://demo-api.carledgr.com/api
    description: Demo server
  - url: https://api.carledgr.com/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Get organization users (Owner only)
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    post:
      tags: [Users]
      summary: Create organization user (Owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created
...
  /users/{id}:
    put:
      tags: [Users]
      summary: Update organization user (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: User updated

    delete:
      tags: [Users]
      summary: Delete organization user (Owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted

  # Organization Management
...
  /organizations:
    get:
      tags: [Organizations]
      summary: Get all organizations (Owner only)
      responses:
        '200':
          description: List of organizations

    post:
      tags: [Organizations]
      summary: Create organization (Owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganization'
      responses:
        '201':
          description: Organization created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [owner, manager, operator]
        roleId:
          type: integer
        organizationId:
          type: string
          format: uuid
        organization_name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUser:
      type: object
      required: [email, password, first_name, last_name, role_id]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        first_name:
          type: string
        last_name:
          type: string
        role_id:
          type: integer
          minimum: 1
          maximum: 3
    UpdateUser:
      type: object
      required: [email, first_name, last_name, role_id]
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role_id:
          type: integer
          minimum: 1
          maximum: 3
...
    CreateOrganization:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
...